requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.5.0
pydantic==2.6.1
sqlalchemy==2.0.27
python-dotenv==1.0.0
httpx==0.27.0
pytest==7.4.4
pytest-asyncio==0.21.2
pytest-cov==4.1.0 